move lexer out of main.rs
